refactor: add new class ProgrammaticOrdersApi
